compact integration of spring, resteasy, and jackson2
This project illustrate the only setup that I can find that reconsiles resteasy, spring, and providers in a straightforward and easy to understand way. Resteasy resources and providers are made to be spring components in addition to being resteasy components. Since resteasy resources are also spring components, dependency injecting them with spring business component works as well as any regular spring component.

1. First resteasy registers its BeanPostProcessor in spring and fires the spring context's loader.

2. Second, spring loads all the beans it is supposed to manage and in particular resteasy components. In this example, spring is set to scan for components sitting in a designated package.

3. Then, the resteasy BeanPostProcessor that was regsitred is fired by spring to scans the spring context for components having the jax-rs annotations like @Provider, @Path...
This means that a provider must be at the same time a spring component decorated with the @Component annotation and a resteasy component decorated with the @Provider annotation. The same goes for a resource by decorating it respectively with the @Component annotation and the @Path annotation.

4. The resteasy resources are wired to the resteasy servlet to receive incoming requests.

5. Providers are registred in the resteasy and used to marshall/unmarshall objects