using servlet 3.0 to simplify resteasy setup
resteasy makes it possible to register its servlet in java using the servlet 3.0 container capabilities. The org.jboss.resteasy:resteasy-servlet-initializer contains a ServletContainerInitializer that is programmed to scan the classpath lookingn for a javax.ws.rs.core.Application decorated with the @javax.ws.rs.ApplicationPath annotation and uses the configuration provides in this class to setup the rest servlet.

The difference is instead of configuring a servlet inside